Busqueda de clientes en tabla, menu de exportacion

+++ frontend/src/components/clientes/TablaClientes.jsx
@@ -3,15 +3,40 @@ import useSocket from "../../hooks/useSocket"; // Importa el hook
 import clienteAxios from "../../config/axios";
 import PropTypes from "prop-types";
 
+
 import { HiSearchCircle } from "react-icons/hi";
 import { RiFileExcel2Fill } from "react-icons/ri";
 import { FaFilter, FaUserEdit } from "react-icons/fa";
 import { MdDelete } from "react-icons/md";
+import ConfirmDialogDelete from "../ui/confirmDialogDelete";
+import MenuExport from "../ui/MenuExport";
 
 const TablaClientes = ({ openModal }) => {
   const [clientes, setClientes] = useState([]);
   const [loading, setLoading] = useState(true);
   const [deleteCliente, setDeleteCliente] = useState(null);
+  const [searchTerm, setSearchTerm] = useState("");
+  const [optionsExport, setOptionsExport] = useState(null);
+
+ 
+  const filteredClientes = clientes.filter((cliente) =>
+    `${cliente.nombre} ${cliente.apellido_paterno} ${cliente.apellido_materno} ${cliente.matricula} ${cliente.telefono}`
+      .toLowerCase()
+      .includes(searchTerm.toLowerCase())
+  );
+  const toggleOptionsExport = () => {
+    setOptionsExport((prev) => !prev);
+  };
+
+  const handleDownload = () => {
+    console.log("Exportando...");
+    setOptionsExport(false);
+  };
+
+  const handleSendReport = () => {
+    console.log("Enviando reporte...");
+    setOptionsExport(false);
+  };
 
   const handleNewClient = (nuevoCliente) => {
     setClientes((prevClientes) => {
@@ -64,28 +89,53 @@ const TablaClientes = ({ openModal }) => {
 
   return (
     <div className="my-4 p-3 bg-white rounded-lg">
-      <div className="p-2 grid grid-cols-4 gap-5">
-        <h1 className="p-2 font-bold text-xl">Todos los clientes</h1>
-        <div className="col-span-2">
+      <div className="p-2 grid md:grid-cols-4 md:gap-5">
+        <div className=" grid grid-cols-3 md:flex  justify-between">
+          <h1 className="col-span-2 p-2 font-bold text-xl">
+            Todos los clientes
+          </h1>
+          <div className="flex md:hidden w-full md:w-0 items-center">
+            <button className="w-full p-1 hover:bg-slate-900 hover:bg-opacity-25 hover:scale-125 transition-all duration-300">
+              <RiFileExcel2Fill className="m-auto text-2xl " />
+            </button>
+            <button className="w-full p-1 hover:bg-slate-900 hover:bg-opacity-25 hover:scale-125 transition-all duration-300">
+              <FaFilter className="m-auto text-2xl" />
+            </button>
+          </div>
+        </div>
+        <div className="md:col-span-2">
           <form className="flex">
             <input
               type="text"
               placeholder="Buscar usuario..."
+              value={searchTerm}
+              onChange={(e) => setSearchTerm(e.target.value)}
               className="w-full px-4 py-2 border border-gray-300 rounded-2xl shadow-sm focus:outline-none focus:ring-1 focus:ring-zinc-600 focus:border-zinc-800"
             />
             <button
-              type="submit"
+              type="button"
               className=" inset-y-0 right-0 flex -ml-5 items-center px-4 text-white bg-zinc-700 rounded-r-lg hover:bg-zinc-800 focus:ring-2 focus:ring-zinc-300"
             >
               <HiSearchCircle className="text-2xl" />
             </button>
           </form>
         </div>
-        <div className="grid grid-cols-2 md:flex h-auto items-center text-lg">
-          <button className="w-full gap-2 px-3 border-r-2 border-gray-900 flex  justify-center items-center hover:bg-zinc-600 hover:bg-opacity-20">
+        <div className="hidden md:flex divide-x-4 h-auto items-center text-lg">
+          <button
+            onClick={toggleOptionsExport}
+            type="button"
+            className="scale-hover-10 w-full gap-2 px-3  flex  justify-center items-center hover:bg-zinc-600 hover:bg-opacity-20"
+          >
             <RiFileExcel2Fill /> Exportar
           </button>
-          <button className="w-full gap-2 px-3 flex  justify-center items-center hover:bg-zinc-600 hover:bg-opacity-20">
+          {/* Recuadro con opciones de exportación */}
+          {optionsExport && (
+             <MenuExport
+             onDownload={handleDownload}
+             onSendReport={handleSendReport}
+           />
+          )}
+          <button className="scale-hover-10 w-full gap-2 px-3 flex  justify-center items-center hover:bg-zinc-600 hover:bg-opacity-20">
             <FaFilter />
             Filtro
           </button>
@@ -98,6 +148,7 @@ const TablaClientes = ({ openModal }) => {
               <th className="px-5 py-2 text-gray-700 uppercase">#</th>
               <th className="px-5 py-2 text-gray-700 uppercase">Foto</th>
               <th className="px-5 py-2 text-gray-700 uppercase">Nombre</th>
+              <th className="px-5 py-2 text-gray-700 uppercase">Estado</th>
               <th className="px-5 py-2 text-gray-700 uppercase">Matricula</th>
               <th className="px-5 py-2 text-gray-700 uppercase">Telefono</th>
               <th className="px-5 py-2 text-gray-700 uppercase">Email</th>
@@ -105,7 +156,7 @@ const TablaClientes = ({ openModal }) => {
             </tr>
           </thead>
           <tbody className="bg-white divide-y divide-gray-200 text-center items-center">
-            {clientes.map((cliente, index) => (
+            {filteredClientes.map((cliente, index) => (
               <tr
                 key={cliente.id}
                 className={`hover:bg-gray-100 ${
@@ -135,8 +186,12 @@ const TablaClientes = ({ openModal }) => {
                   {cliente.apellido_materno}
                 </td>
                 <td className="px-6 py-4 text-sm font-semibold text-gray-700">
+                  {cliente.estado}
+                </td>
+                <td className="px-6 py-4 text-sm font-semibold text-gray-700">
                   {cliente.matricula}
                 </td>
+
                 <td className="px-6 py-4 text-sm font-semibold text-gray-700">
                   {cliente.telefono}
                 </td>
@@ -162,29 +217,11 @@ const TablaClientes = ({ openModal }) => {
                       <MdDelete className="text-3xl" />
                     </button>
                     {deleteCliente === cliente.id && (
-                      <div className="absolute mt-2 bg-white border rounded-lg shadow-lg w-56 z-10 p-4 ">
-                        <h1 className="text-lg font-semibold text-gray-700 mb-2">
-                          ¿Seguro que deseas eliminar a <br />
-                          <span className="font-bold">
-                            {cliente.nombre} {cliente.apellidoPaterno}
-                          </span>
-                          ?
-                        </h1>
-                        <div className="flex justify-between">
-                          <button
-                            className="bg-gray-200 text-gray-700 py-1 px-3 rounded hover:bg-gray-300"
-                            onClick={() => setDeleteCliente(null)}
-                          >
-                            Cancelar
-                          </button>
-                          <button
-                            className="bg-rose-500 text-white py-1 px-3 rounded hover:bg-rose-600"
-                            onClick={() => handleDelete(cliente.id)}
-                          >
-                            Eliminar
-                          </button>
-                        </div>
-                      </div>
+                      <ConfirmDialogDelete
+                        message={`¿Seguro que deseas eliminar a ${cliente.nombre} ${cliente.apellido_paterno}?`}
+                        onCancel={() => setDeleteCliente(null)}
+                        onConfirm={() => handleDelete(cliente.id)}
+                      />
                     )}
                   </div>
                 </td>
